+++ zippick/src/main/java/zippick/domain/product/dto/response/ProductResponse.java
@@ -6,6 +6,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import zippick.domain.product.dto.ProductDto;
 
 @Getter
 @Builder
