+++ zippick/src/main/java/zippick/domain/product/service/ProductServiceImpl.java
@@ -7,7 +7,7 @@ import org.json.JSONArray;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 import zippick.domain.product.dto.response.ProductDetailResponse;
-import zippick.domain.product.dto.response.ProductDto;
+import zippick.domain.product.dto.ProductDto;
 import zippick.domain.product.dto.response.ProductResponse;
 import zippick.domain.product.mapper.ProductMapper;
 import lombok.RequiredArgsConstructor;
